requirements.txt
Django==5.1.7
Pillow==11.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.6.0
reportlab==4.2.5
